Refactor: DiscountCondition 클래스
