+++ app/src/main/java/com/orchsik/object/_05_movie/PeriodCondition.java
@@ -1,0 +1,17 @@
+package com.orchsik.object._05_movie;
+
+import java.time.DayOfWeek;
+import java.time.LocalTime;
+
+public class PeriodCondition implements DiscountCondition {
+  private DayOfWeek dayOfWeek;
+  private LocalTime startTime;
+  private LocalTime endTime;
+
+  public boolean isSatisfiedBy(Screening screening) {
+    return dayOfWeek.equals(screening.getWhenScreened().getDayOfWeek())
+        && startTime.compareTo(screening.getWhenScreened().toLocalTime()) <= 0
+        && endTime.compareTo(screening.getWhenScreened().toLocalTime()) >= 0;
+  }
+
+}
