+++ app/src/main/java/com/orchsik/object/_05_movie/SequenceCondition.java
@@ -1,0 +1,12 @@
+package com.orchsik.object._05_movie;
+
+public class SequenceCondition implements DiscountCondition {
+
+  private int sequence;
+
+  public boolean isSatisfiedBy(Screening screening) {
+    return sequence == screening.getSequence();
+
+  }
+
+}
